BUG: 215476 CCMAIL: kde-core-devel@kde.org
Introduce a configuration option that disables the left-mouse-button
viewport dragging. I have been pinged multiple times now that some
people find it annoying.

I know that Palapeli is in kdereview (therefore I'm CCing k-c-d)
currently, but feature freeze is very near, and I want this easy change
in before it, esp. because it introduces new strings.

I'm unsure about the default value of this configuration option, though.
I think I'll make a poll for the default setting on my blog when the
Beta 1 is out and Palapeli becomes available to more testers.


svn path=/trunk/kdereview/palapeli/; revision=1052165