Massive performance optimization of cursor management for big puzzles.
Call QGraphicsItem::setCursor for the Part as a whole rather than for
the single pieces. This means that the cursor is now also set for shadow
items. I have consequently enabled mouse interaction for shadow items,
too, in order to have consistent behavior.

This in turn "fixes" feature request 212842, but it introduces an
interesting bug: When one clicks on a piece that is covered by another
piece's shadow, the other piece will be selected instead. I consider
this bug less important than the above-named performance problems of
QGraphicsItem::setCursor.

CCBUG: 212842

svn path=/trunk/KDE/kdegames/palapeli/; revision=1062639