Update the puzzles for KDE 4.4 RC 3.
I have to manually address an issue which I forgot when designing the
puzzle file format: The puzzle files contain translatable strings in
such a way that they are not visible to scripty. They have to be built
by hand to update the translations. This issue will be fixed with
Palapeli 1.1 (SC 4.5), but I cannot do this in 1.0 because I have to
change quite a bit of the puzzle file loader for that.


svn path=/branches/KDE/4.4/kdegames/palapeli/; revision=1081401