Begin to redesign the core engine for better performance on big images.
Some functions of Part and ShadowItem are absorbed in Piece, and the
internal data structures change massively. It will become clear in the
next commits what I'm doing. The code for the old Part and Piece are
left in the source tree as a quick reference.

This port (temporarily) breaks piece merging and loading/saving.

svn path=/trunk/KDE/kdegames/palapeli/; revision=1111383