Add highlighting scheme for when shadows are unavailable or hard to see.
Previously in Palapeli, selection highlighting has been based on
changing the color of the piece shadows. This breaks down if the shadow
setting is off, the highlighting color is dim (as on Apple Mac) or the
pieces are very small (as in a full view of a puzzle table with 1000s
of pieces). Shadowing costs puzzle load time and may often be disabled.