Refactor parts of Palapeli::Scene into GamePlay: first working version.
Methods relating to the puzzle table as a whole have been moved from
Palapeli::Scene into Palapeli::GamePlay, including puzzle loading and
shuffling, puzzle saving and restoring, victory animation and progress
reporting. This is to provide a simpler, more generic Palapeli::Scene
class that can be used in multiple scenes, i.e. the original puzzle
table and any number of piece-holders for use in solving large puzzles.

The main methods remaining in Palapeli::Scene are adding and removing
pieces, moving pieces, merging (or joining) pieces and signalling
changes in the state of the puzzle and its pieces.