Make ConstraintVisualizer thinner and within a margin around all pieces.
Previously the draggable frame (ConstraintVisualizer) around the puzzle
table was wide and could cover entire pieces, which butted right up to
the edge of the frame. This made it harder to see and select pieces.
Also, to avoid an 'ugly' look, the puzzle would start with empty space
around the frame.

Now, with a margin and a thinner frame, the 'ugly' look can be avoided
and the puzzle can fill the view when it starts.