Add code for teleporting pieces between holders and puzzle table.
This code is driven by selecting pieces and holders and by using a
special mouse-click (default Shift-LeftClick) to trigger teleporting
(instant transfer of pieces) from puzzle table to piece-holder, holder
to puzzle table and holder to holder. It allows pieces to be collected
and sorted quickly when searching through the puzzle-table of large
puzzles (300 to 10,000 pieces).