Make sure that no interaction is active before clearing the scene
Summary:
When the scene/views are cleared and there is still an interaction active, palapeli will crash.
This happens to due the fact that the interactors are holding pointers to pieces which are cleared
by Scene::clearPieces()
BUG 336075

Test Plan:
As described in the bug description: Create a shortcut for restarting the puzzle, select a piece
and move it around. In the meantime restart the puzzle -> no more crash

Reviewers: #kde_games

Subscribers: #kde_games

Differential Revision: https://phabricator.kde.org/D7637