Add an owner scene to each piece, and merge only pieces with the same owner
This seems to solve a crash that sometimes occured when using piece holders.
One of the symptoms was that when trying to call removeAll on one of the
pieces to be removed, the return value from removeAll was zero, indicating that
no piece was removed from the list - so it must have been in a different list.
The code for merging pieces seems to have been unaware of the existence of
multiple scenes.