Make the cache work
There were two problems here: the QDateTime comparisons didn't return
true for files with unchanged mtime.  Using the same conversion on both
sides of the comparison fixes that.  As a result, startup now becomes
much faster since we're not rebuilding the cache each time.  Since
the format of the date has changed on-disk, this should be backwards
compatible, causing a one-time rebuild.

That exposed another problem that I was half-observing already before,
which is that the PNG thumbnails weren't being restored properly.
Converting the raw data to base64 when saving fixes that as well.