First pass of removing threading
My guess would be that this was an optimization for very slow load times,
which I've also seen.  But the reason for these are repeated fdatasync
calls which are completely unnecessary. By removing the threading, we can
actually more easily optimize this in a followup patch.

This is probably not a complete removal of all the threading-related cruft.
I suspected a number of mutexes are no longer necessary now, but that can
be left for later.