Port slicer plugin loading to KPluginLoader
Loading plugins via KServiceTypeTrader is deprecated. Instead the
plugins should embed their metadata, be installed in a well-known
subdirectory of the plugin dir and loaded via
KPluginLoader/KPluginMetaData.

Any existing out-of-tree plugin would need to be adapted, otherwise it
will not be loaded. The adjustment is easy however and I doubt there are
actually any out-of-tree plugins.